@@ -3,20 +3,19 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@material-ui/core": "^4.11.3",
-    "@material-ui/icons": "^4.11.2",
+    "@dnd-kit/core": "^6.3.1",
+    "@emotion/react": "^11.14.0",
+    "@emotion/styled": "^11.14.1",
+    "@mui/icons-material": "^7.3.5",
+    "@mui/material": "^7.3.5",
     "lodash": "^4.17.20",
-    "react": "^16.14.0",
-    "react-animated-slider": "^2.0.0",
-    "react-canvas-draw": "^1.1.1",
-    "react-dnd": "^11.1.3",
-    "react-dnd-html5-backend": "^11.1.3",
-    "react-dnd-multi-backend": "^6.0.2",
-    "react-dnd-touch-backend": "^11.1.3",
-    "react-dom": "^16.14.0",
+    "react": "^19.2.0",
+    "react-dom": "^19.2.0",
+    "react-konva": "^19.2.0",
     "react-player": "^2.8.2",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "^3.4.4"
+    "react-router-dom": "^7.9.5",
+    "react-scripts": "^5.0.1",
+    "swiper": "^12.0.3"
   },
   "scripts": {
     "start": "react-scripts start",
