@@ -1,24 +1,8 @@
 import React from 'react';
 import { alphabets } from '../utils/alphabets';
-import { Grid, Button, withStyles, Typography } from '@material-ui/core';
-import { Replay, NavigateNext, Forward } from '@material-ui/icons';
+import { Grid, Button, Typography } from '@mui/material';
 
-const styles = () => ({
-  buttonStyle: {
-    margin: '4px 0px 0px 4px',
-  },
-  restartButton: {
-    backgroundColor: 'blue'
-  },
-  nextButton: {
-    backgroundColor: 'green',
-    color: 'white',
-    '&:hover': {
-      backgroundColor: '#9acd32',
-      color: 'blue',
-    }
-  },
-});
+import { Replay, NavigateNext, Forward } from '@mui/icons-material';
 
 class ShortVowelTest extends React.Component{
   constructor(props) {
@@ -87,29 +71,29 @@ class ShortVowelTest extends React.Component{
   }
   render() {
     const { learnedLocalAlphabets, displayCurrent, displayVowel } = this.state;
-
-    const { classes } = this.props;
     
-    return <React.Fragment>
-      <Grid container justify="center" style={{border: '1px solid #224422', padding: '5px'}}>
-        <Grid item>
-          <Button className={`${classes.buttonStyle} ${classes.restartButton}`}
-            variant="contained" color="primary" onClick={this.restart}
-            startIcon={<Replay />}
-            >Restart</Button>
-          <Button
-            className={`${classes.buttonStyle} ${classes.nextButton}`}
-            startIcon={<NavigateNext />}
-            endIcon={<Forward />}
-            onClick={this.validateAnswer}
-            >Next</Button>
+    return (
+      <React.Fragment>
+        <Grid container justifyContent="center" style={{border: '1px solid #224422', padding: '5px'}}>
+          <Grid item>
+            <Button className="buttonStyle restartButton"
+              variant="contained" color="primary" onClick={this.restart}
+              startIcon={<Replay />}
+              >Restart</Button>
+            <Button
+              className="buttonStyle nextButton"
+              startIcon={<NavigateNext />}
+              endIcon={<Forward />}
+              onClick={this.validateAnswer}
+              >Next</Button>
+          </Grid>
         </Grid>
-      </Grid>
-        <Typography component="div"
-            style={{textAlign: 'center', color: 'blue', padding: '10px', fontSize: '6em'}}>
-              {learnedLocalAlphabets[displayCurrent].harakat.letters.split(' ')[displayVowel]}</Typography>
-    </React.Fragment>;
+          <Typography component="div"
+              style={{textAlign: 'center', color: 'blue', padding: '10px', fontSize: '6em'}}>
+                {learnedLocalAlphabets[displayCurrent].harakat.letters.split(' ')[displayVowel]}</Typography>
+      </React.Fragment>
+    );
   }
 }
 
-export default withStyles(styles)(ShortVowelTest);
+export default ShortVowelTest;
