@@ -1,21 +1,20 @@
 import React from 'react';
-import { withStyles, FormControlLabel, Radio,
-  RadioGroup } from '@material-ui/core';
-const styles = () => ({
+import { FormControlLabel, Radio, RadioGroup } from '@mui/material';
+
+const styles = {
   label: {
     fontSize: '1em',
-  },
-});
+  }};
 
 function AnswerOptions(props){
-  const { classes, options, handleAnswerOptionClick, optionValue, optionLabel } = props;
+  const { options, handleAnswerOptionClick, optionValue, optionLabel } = props;
   
   return <React.Fragment>
     <RadioGroup row style={{ justifyContent: 'space-between', padding: '10px' }}>
       {options && options.map(opt => <FormControlLabel
         key={optionValue ? opt[optionValue] : opt}
         onChange={handleAnswerOptionClick}
-        classes={{ label: classes.label }}
+        classes={{ label: styles.label }}
         control={<Radio />}
         value={optionValue ? opt[optionValue] : opt}
         label={optionLabel ? opt[optionLabel] : opt} />)}
@@ -23,4 +22,4 @@ function AnswerOptions(props){
   </React.Fragment>;
 }
 
-export default withStyles(styles)(AnswerOptions);
+export default AnswerOptions;
