@@ -1,55 +1,21 @@
 import React from 'react';
-import { Button, withStyles, Typography, Dialog,
-  DialogTitle, DialogContent, DialogActions, DialogContentText, } from '@material-ui/core';
-import { EmojiEmotions, MoodBad } from '@material-ui/icons';
+import {
+  Button,
+  Typography,
+  Dialog,
+  DialogTitle,
+  DialogContent,
+  DialogActions,
+  DialogContentText,
+} from '@mui/material';
+
+import { EmojiEmotions, MoodBad } from '@mui/icons-material';
 import _ from 'lodash';
 import TopPanel from '../common/Exercise/TopPanel';
 import Question from '../common/Exercise/Question';
 import AnswerOptions from '../common/Exercise/AnswerOptions';
 import { QUESTION_TYPES, generateRandomNumber, pickRandomFromArray, generateQuestionOptions } from '../../utils/utils';
 import { arabicMonths } from '../../utils/months';
-
-const styles = () => ({
-  label: {
-    fontSize: '3em',
-  },
-  buttonStyle: {
-    margin: '4px 0px 0px 4px',
-  },
-  restartButton: {
-    backgroundColor: 'blue'
-  },
-  nextButton: {
-    backgroundColor: 'green',
-    color: 'white',
-    '&:hover': {
-      backgroundColor: '#9acd32',
-      color: 'blue',
-    }
-  },
-  questNumber: {
-    fontSize: '1.5em',
-    fontWeight: 'bold',
-  },
-  sliderValueLabel: {
-    color: 'green',
-  },
-  rail: {
-    backgroundColor: 'red',
-    height: '15px',
-  },
-  track: {
-    backgroundColor: 'green',
-    height: '15px',
-  },
-  result: {
-    fontSize: '1.5em'
-  },
-  resultHighlight: {
-    fontSize: '2em',
-    color: 'blue'
-  }
-});
 
 class MonthExerciseTwo extends React.Component{
   constructor(props) {
@@ -161,7 +127,6 @@ class MonthExerciseTwo extends React.Component{
   render() {
     const { openResultAlert, questionNumber, displayFinalResult, maxQuestions, questionWord,
       right, wrong, answerOptions, result, questionText, optionLabel, optionValue, answer } = this.state;
-    const { classes } = this.props;
     return <React.Fragment>
       <div className="instruction"> Choose the right answer from the options.</div>
       <TopPanel
@@ -180,7 +145,7 @@ class MonthExerciseTwo extends React.Component{
         options={answerOptions}
         optionLabel={optionLabel}
         optionValue={optionValue}
-        classes={{label: optionLabel === 'arabic' ? classes.label : null}}
+        classes={{label: optionLabel === 'arabic' ? {fontSize: '24px'} : null}}
         handleAnswerOptionClick={this.handleAnswerOptionClick}
         />
       <Dialog
@@ -201,10 +166,10 @@ class MonthExerciseTwo extends React.Component{
                 </Typography>}
         </DialogTitle>
         <DialogContent>
-          <DialogContentText id="alert-dialog-description" className={classes.result}>
-            The correct answer is <span className={classes.resultHighlight}> {_.get(questionWord, [optionLabel])}
+          <DialogContentText id="alert-dialog-description" className="result">
+            The correct answer is <span className="resultHighlight"> {_.get(questionWord, [optionLabel])}
             </span><br />
-            You have selected <span className={classes.resultHighlight}> {answer} </span>
+            You have selected <span className="resultHighlight"> {answer} </span>
           </DialogContentText>
         </DialogContent>
         <DialogActions>
@@ -219,4 +184,4 @@ class MonthExerciseTwo extends React.Component{
   }
 }
 
-export default withStyles(styles)(MonthExerciseTwo);
+export default MonthExerciseTwo;
