@@ -1,16 +1,17 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
 import './index.css';
+import ReactDOM from "react-dom/client";
 import * as serviceWorker from './serviceWorker';
-import { HashRouter } from 'react-router-dom';
+import { BrowserRouter } from 'react-router-dom';
 import App from './App';
 
-ReactDOM.render(
+
+
+const root = ReactDOM.createRoot(document.getElementById('root'));
+    root.render(
   <React.StrictMode>
-    <HashRouter>
+    <BrowserRouter>
       <App />
-    </HashRouter>
-  </React.StrictMode>,
-  document.getElementById('root')
-);
+    </BrowserRouter>
+  </React.StrictMode>);
 serviceWorker.unregister();
