@@ -1,7 +1,4 @@
 import React from 'react';
-import CanvasDraw from 'react-canvas-draw';
-import { Button, Hidden } from '@material-ui/core';
-import { Backspace } from '@material-ui/icons';
 
 class FunLevelOne extends React.Component {
   constructor(props) {
@@ -20,44 +17,45 @@ class FunLevelOne extends React.Component {
     w.document.write("<img src='"+ d +"' style='background-image: url( " + img + "); background-size: contain' alt='Exporting'/>");
   }
   render() {
-    const { brushColor } = this.state;
-    return <React.Fragment>
-      <div style={{display: 'flex', justifyContent: 'center', overflow:'scroll'}}>
-        <div style={{border: '2px solid #4f111d'}}><Hidden mdUp>
-          <Button variant="contained" color="primary"
-            onClick={this.eraseWindow}
-            startIcon={<Backspace />}
-            >Erase All</Button>
-          <CanvasDraw
-            ref={canvasDraw => (this.saveableCanvas = canvasDraw)}
-            lazyRadius={0}
-            brushColor={brushColor}
-            brushRadius={3}
-            imgSrc={require("../images/fun/fun-rabbit1.png")}
-            canvasWidth="350px"
-            canvasHeight="330px"
-            catenaryColor={brushColor}
-            />
-        </Hidden>
-        <Hidden smDown>
-          <Button variant="contained" color="primary"
-            onClick={this.eraseWindow}
-            startIcon={<Backspace />}
-            >Erase All</Button>
-          <CanvasDraw
-            ref={canvasDraw => (this.saveableCanvas = canvasDraw)}
-            lazyRadius={0}
-            brushColor={brushColor}
-            brushRadius={3}
-            imgSrc={require("../images/fun/fun-rabbit1.png")}
-            canvasWidth="560px"
-            canvasHeight="530px"
-            catenaryColor={brushColor}
-            />
-        </Hidden>
+    return (
+      <React.Fragment>
+        <div style={{display: 'flex', justifyContent: 'center', overflow:'scroll'}}>
+          <div style={{border: '2px solid #4f111d'}}>{/* <Hidden mdUp>
+            <Button variant="contained" color="primary"
+              onClick={this.eraseWindow}
+              startIcon={<Backspace />}
+              >Erase All</Button>
+            {/* <CanvasDraw
+              ref={canvasDraw => (this.saveableCanvas = canvasDraw)}
+              lazyRadius={0}
+              brushColor={brushColor}
+              brushRadius={3}
+              imgSrc={require("../images/fun/fun-rabbit1.png")}
+              canvasWidth="350px"
+              canvasHeight="330px"
+              catenaryColor={brushColor}
+              /> 
+          </Hidden>
+          <Hidden mdDown>
+            <Button variant="contained" color="primary"
+              onClick={this.eraseWindow}
+              startIcon={<Backspace />}
+              >Erase All</Button>
+            {/* <CanvasDraw
+              ref={canvasDraw => (this.saveableCanvas = canvasDraw)}
+              lazyRadius={0}
+              brushColor={brushColor}
+              brushRadius={3}
+              imgSrc={require("../images/fun/fun-rabbit1.png")}
+              canvasWidth="560px"
+              canvasHeight="530px"
+              catenaryColor={brushColor}
+              /> 
+          </Hidden> */}
+          </div>
         </div>
-      </div>
-    </React.Fragment>
+      </React.Fragment>
+    );
   }
 }
 
